refacto: move the snack bar to the scaffold subpackage
