+++ app/src/main/java/com/illiouchine/jm/ui/screen/PollSetupScreen.kt
@@ -31,7 +31,7 @@ import com.illiouchine.jm.R
 import com.illiouchine.jm.logic.PollSetupViewModel
 import com.illiouchine.jm.model.Grading
 import com.illiouchine.jm.ui.composable.GradingSelectionRow
-import com.illiouchine.jm.ui.composable.MjuSnackbarWithStringResId
+import com.illiouchine.jm.ui.composable.scaffold.MjuSnackbarWithStringResId
 import com.illiouchine.jm.ui.composable.ProposalRow
 import com.illiouchine.jm.ui.composable.ProposalSelectionRow
 import com.illiouchine.jm.ui.composable.ScreenTitle
